+++ Calculator/Mathematics/Operations/Addition.cs
@@ -6,7 +6,7 @@ namespace Calculator.Mathematics.Operations
     {
         public char Operator { get; set; } = '+';
 
-        public decimal Execute(decimal first, decimal second)
+        public double Execute(double first, double second)
         {
             return Math.Round(first + second, 2);
         }
