+++ Database/Models/MathCalculation.cs
@@ -12,10 +12,10 @@ namespace Database.Models
         public int Id { get; set; }
         public char Operator { get; set; }
 
-        public decimal Answer { get; set; }
+        public double Answer { get; set; }
 
-        public decimal FirstInput { get; set; }
-        public decimal SecondInput { get; set; }
+        public double FirstInput { get; set; }
+        public double SecondInput { get; set; }
 
         [Required]
         public Result Result { get; set; } = null!;
